# lib/api-spec/openapi.yaml
...
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AURELION - Luxury Aruba Adventure Planning Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /itineraries:
    get:
      operationId: listItineraries
      tags: [itineraries]
      summary: List user itineraries
      responses:
        "200":
          description: List of itineraries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Itinerary"
    post:
      operationId: createItinerary
      tags: [itineraries]
      summary: Create itinerary
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateItineraryBody"
      responses:
        "201":
          description: Itinerary created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Itinerary"
  /itineraries/{id}:
    get:
      operationId: getItinerary
      tags: [itineraries]
      summary: Get itinerary detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Itinerary with items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItineraryDetail"
        "404":
          description: Not found
    patch:
      operationId: updateItinerary
      tags: [itineraries]
      summary: Update itinerary
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateItineraryBody"
      responses:
        "200":
          description: Updated itinerary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Itinerary"
    delete:
      operationId: deleteItinerary
      tags: [itineraries]
      summary: Delete itinerary
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /itineraries/{id}/items:
    post:
      operationId: addItineraryItem
      tags: [itineraries]
      summary: Add activity to itinerary
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddItineraryItemBody"
      responses:
        "201":
          description: Item added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItineraryItem"
  /itineraries/{id}/items/{itemId}:
    patch:
      operationId: updateItineraryItem
      tags: [itineraries]
      summary: Update itinerary item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateItineraryItemBody"
      responses:
        "200":
          description: Updated item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItineraryItem"
    delete:
      operationId: removeItineraryItem
      tags: [itineraries]
      summary: Remove item from itinerary
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
  /itineraries/{id}/export:
    get:
      operationId: exportItinerary
      tags: [itineraries]
      summary: Export itinerary (requires purchase)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Exported itinerary data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItineraryExport"
        "403":
          description: Tier upgrade required
...
components:
  schemas:
...
    Activity:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        category:
          type: string
        difficulty:
          type: string
        durationMinutes:
          type: integer
        priceLow:
          type: number
        priceHigh:
          type: number
        location:
          type: string
        imageUrl:
          type: ["string", "null"]
        reviewSummary:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        description:
          type: string
        whatToBring:
          type: ["string", "null"]
        whatToExpect:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - title
        - category
        - difficulty
        - durationMinutes
        - priceLow
        - priceHigh
        - location
        - description
        - tags
        - createdAt
...
    Itinerary:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        totalDays:
          type: integer
        tierType:
          type: string
        status:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - userId
        - title
        - totalDays
        - tierType
        - status
        - createdAt
        - updatedAt
    ItineraryDetail:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        totalDays:
          type: integer
        tierType:
          type: string
        status:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/ItineraryItemWithActivity"
      required:
        - id
        - userId
        - title
        - totalDays
        - tierType
        - status
        - createdAt
        - updatedAt
        - items
    ItineraryItem:
      type: object
      properties:
        id:
          type: integer
        itineraryId:
          type: integer
        activityId:
          type: integer
        dayNumber:
          type: integer
        timeSlot:
          type: string
        notes:
          type: ["string", "null"]
      required:
        - id
        - itineraryId
        - activityId
        - dayNumber
        - timeSlot
    ItineraryItemWithActivity:
      type: object
      properties:
        id:
          type: integer
        itineraryId:
          type: integer
        activityId:
          type: integer
        dayNumber:
          type: integer
        timeSlot:
          type: string
        notes:
          type: ["string", "null"]
        activity:
          $ref: "#/components/schemas/Activity"
      required:
        - id
        - itineraryId
        - activityId
        - dayNumber
        - timeSlot
        - activity
    CreateItineraryBody:
      type: object
      properties:
        title:
          type: string
        totalDays:
          type: integer
      required:
        - title
        - totalDays
    UpdateItineraryBody:
      type: object
      properties:
        title:
          type: string
        totalDays:
          type: integer
        status:
          type: string
    AddItineraryItemBody:
      type: object
      properties:
        activityId:
          type: integer
        dayNumber:
          type: integer
        timeSlot:
          type: string
        notes:
          type: string
      required:
        - activityId
        - dayNumber
        - timeSlot
    UpdateItineraryItemBody:
      type: object
      properties:
        dayNumber:
          type: integer
        timeSlot:
          type: string
        notes:
          type: ["string", "null"]
    ItineraryExport:
      type: object
      properties:
        itinerary:
          $ref: "#/components/schemas/ItineraryDetail"
        tierType:
          type: string
        exportedAt:
          type: string
      required:
        - itinerary
        - tierType
        - exportedAt
...